+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Development
+  - Provide additional trace logging for distributed key generation
   - Enforce requirement that peer names cannot be the same
   - Exit early in distributed key generation if execution fails
   - Allow `process.generation-passphrase` to be a Majordomo URL